Initial Commit of Game Logic
This is my first Commit pertaining to the Tic Tac Toe game logic.  There
are two type of elements implemented in this game logic:
    1) Input processing and game rules
    2) Visual feedback

INPUT PROCESSING AND GAME RULES
My game currently detects wins, ties, and whether a move is valid.
The game also tracks the number of wins and whose turn it currently is.
The game also allows you to restart the game.

VISUAL FEEDBACK
For visual feedback, I included JQuery that helps message information
to the player.
This includes alerting the player of whose turn it currently is,
identifying who won, or alerting the players if it is a tie game.
This commit also contains a visual score tracker.
Data is not currently being pushed or stored in the backend server.

FUTURE ADDITIONS
Future additions will include api methods to store the data.